requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-WTF==1.1.1
SQLAlchemy==2.0.21
itsdangerous==2.1.2
websockets==11.0.3
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
Pillow==10.0.1
pytz==2023.3
python-dotenv==1.0.0
WTForms==3.0.1
pathlib2==2.3.7


